+++ h/toss.h
@@ -63,8 +63,8 @@ void tossFromBadArea(char force);
 void writeMsgToSysop(void);
 int  checkAreaLink(s_area *area, hs_addr aka, int type);
 void forwardMsgToLink(s_message *msg, s_area *echo, s_link *link,
-                    s_seenBy *seenBys, UINT seenByCount,
-                    s_seenBy *pathArray, UINT pathArrayCount);
+                    s_seenBy *seenBys, UINT16 seenByCount,
+                    s_seenBy *pathArray, UINT16 pathArrayCount);
 void forwardMsgToLinks(s_area *echo, s_message *msg, hs_addr pktOrigAddr);
 int  carbonCopy(s_message *msg, XMSG *xmsg, s_area *echo);
 s_arealink *getAreaLink(s_area *area, hs_addr aka);
